more refactoring and pass createPoll action to navigation screens through screenProps
