+++ Polly/src/scenes/CreatePoll/CreatePollContainer.js
@@ -1,0 +1,21 @@
+import React from 'react'
+import CreatePoll from './CreatePoll'
+import { createPoll } from '../../actions/polls'
+import { connect } from 'react-redux'
+
+class CreatePollContainer extends React.Component {
+  render() {
+    console.log(this.props)
+    return (
+      <CreatePoll screenProps={{createPoll: this.props.createPoll}} />
+    )
+  }
+}
+
+const mapDispatchToProps = dispatch => {
+  return {createPoll: pollData => dispatch(createPoll(pollData))}
+}
+
+CreatePollContainer = connect(null, mapDispatchToProps)(CreatePollContainer)
+
+export default CreatePollContainer
