+++ Polly/src/scenes/CreatePoll/CreatePollScreen.js
@@ -3,7 +3,8 @@ import { connect } from 'react-redux'
 import { ThemeProvider, Button } from 'react-native-elements'
 import { View, StyleSheet } from 'react-native'
 import { createPoll } from '../../actions/polls'
-import { theme, elevationShadowStyle } from './index'
+import { theme, elevationShadowStyle } from './style'
+
 
 class CreatePollScreen extends React.Component {
     static navigationOptions = {
@@ -11,28 +12,29 @@ class CreatePollScreen extends React.Component {
     };
     render() {
       const {navigate} = this.props.navigation;
+      console.log(this.props)
       return(
         <ThemeProvider theme={theme}>
           <View style={styles.view}> 
             <Button
               style={styles.container}
               title="True False"
-              onPress={() => navigate('TrueFalsePoll', {name: 'TrueFalse', createPoll: this.props.createPoll})}
+              onPress={() => navigate('TrueFalsePoll', {name: 'TrueFalse', createPoll: this.props.screenProps.createPoll})}
             />
             <Button
               style={styles.container}
               title="Short Answer"
-              onPress={() => navigate('ShortAnswerPoll', {name: 'ShortAnswer', createPoll: this.props.createPoll})}
+              onPress={() => navigate('ShortAnswerPoll', {name: 'ShortAnswer', createPoll: this.props.screenProps.createPoll})}
            />
             <Button
               style={styles.container}
               title="Number Scale"
-              onPress={() => navigate('NumberScalePoll', {name: 'NumberScale', createPoll: this.props.createPoll})}
+              onPress={() => navigate('NumberScalePoll', {name: 'NumberScale', createPoll: this.props.screenProps.createPoll})}
             />
             <Button
               style={styles.container}
               title="Select All"
-              onPress={() => navigate('SelectAllPoll', {name: 'SelectAll', createPoll: this.props.createPoll})}
+              onPress={() => navigate('SelectAllPoll', {name: 'SelectAll', createPoll: this.props.screenProps.createPoll})}
             />
           </View>
         </ThemeProvider>
@@ -52,8 +54,4 @@ const styles = StyleSheet.create({
   }
 });
 
-const mapDispatchToProps = dispatch => {
-  return {createPoll: pollData => dispatch(createPoll(pollData))}
-}
-
-export default connect(null, mapDispatchToProps)(CreatePollScreen)
+export default CreatePollScreen
